[feat]: @DocumentedApiErrors 구현
