+++ src/main/java/com/ggamja/global/config/SecurityConfig.java
@@ -1,4 +1,4 @@
-package com.ggamja.config;
+package com.ggamja.global.config;
 
 import jakarta.servlet.http.HttpServletResponse;
 import org.springframework.context.annotation.Bean;
